Moved Cryptography to encrypted-data-model, fixed integration test
