+++ node/src/main/java/ch/prevo/open/node/services/JobEndService.java
@@ -1,6 +1,7 @@
 package ch.prevo.open.node.services;
 
 import ch.prevo.open.encrypted.model.InsurantInformation;
+import ch.prevo.open.encrypted.services.Cryptography;
 import ch.prevo.open.node.config.AdapterServiceConfiguration;
 import ch.prevo.open.node.data.provider.JobEndProvider;
 import ch.prevo.open.node.data.provider.ProviderFactory;
@@ -18,18 +19,16 @@ import java.util.stream.Collectors;
 public class JobEndService {
 
     private final JobEndProvider jobEndProvider;
-    private final Cryptography cryptography;
 
     @Inject
-    public JobEndService(ServiceListFactoryBean factoryBean, Cryptography cryptography) {
+    public JobEndService(ServiceListFactoryBean factoryBean) {
         final ProviderFactory factory = AdapterServiceConfiguration.getAdapterService(factoryBean);
         this.jobEndProvider = factory != null? factory.getJobEndProvider() : null;
-        this.cryptography = cryptography;
     }
 
 	public Set<InsurantInformation> getAllJobEndData() {
 		return jobEndProvider.getJobEnds().stream()
-				.map(jobEnd -> new InsurantInformation(cryptography.hash(jobEnd.getJobInfo().getOasiNumber()),
+				.map(jobEnd -> new InsurantInformation(Cryptography.hash(jobEnd.getJobInfo().getOasiNumber()),
 						jobEnd.getJobInfo().getRetirementFundUid(), jobEnd.getJobInfo().getDate()))
 				.collect(Collectors.toSet());
 	}
