+++ node/src/test/java/ch/prevo/open/node/api/JobStartControllerTest.java
@@ -2,7 +2,7 @@ package ch.prevo.open.node.api;
 
 import ch.prevo.open.node.NodeApplication;
 import ch.prevo.open.node.data.provider.MockProviderFactory;
-import ch.prevo.open.node.services.Cryptography;
+import ch.prevo.open.encrypted.services.Cryptography;
 import ch.prevo.open.node.services.JobStartService;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -35,19 +35,17 @@ public class JobStartControllerTest extends RestBaseTest {
         public JobStartService jobStartService() throws Exception {
             final ServiceListFactoryBean factory = Mockito.mock(ServiceListFactoryBean.class);
             given(factory.getObject()).willReturn(Collections.singletonList(new MockProviderFactory()));
-            return new JobStartService(factory, new Cryptography());
+            return new JobStartService(factory);
         }
     }
 
     @Test
     public void getAllJobStartData() throws Exception {
-        final Cryptography cryptography = new Cryptography();
-
         mockMvc.perform(get("/job-start"))
                 .andExpect(status().isOk())
                 .andExpect(content().contentType(contentType))
                 .andExpect(jsonPath("$", hasSize(3)))
-                .andExpect(jsonPath("$[0].encryptedOasiNumber", is(cryptography.hash("756.1234.5678.97"))))
+                .andExpect(jsonPath("$[0].encryptedOasiNumber", is(Cryptography.hash("756.1234.5678.97"))))
                 .andExpect(jsonPath("$[0].retirementFundUid", is("CHE-109.740.084")));
     }
 }
