+++ node/src/test/java/ch/prevo/open/node/services/CryptographyTest.java
@@ -1,5 +1,6 @@
 package ch.prevo.open.node.services;
 
+import ch.prevo.open.encrypted.services.Cryptography;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -9,13 +10,12 @@ public class CryptographyTest {
     @Test
     public void hashFunctionShouldIgnoreDots() {
         // given
-        final Cryptography cryptography = new Cryptography();
         final String oasiNumber1 = "756.3412.8844.97";
         final String oasiNumber2 = "7563412884497";
 
         // when
-        final String hash1 = cryptography.hash(oasiNumber1);
-        final String hash2 = cryptography.hash(oasiNumber2);
+        final String hash1 = Cryptography.hash(oasiNumber1);
+        final String hash2 = Cryptography.hash(oasiNumber2);
 
         // then
         assertThat(hash1).isEqualTo(hash2);
